Move ruleset-specific hitwindows to post-converted hitobjects
