+++ osu.Game.Rulesets.Mania/Objects/ManiaHitObject.cs
@@ -1,8 +1,6 @@
 ﻿// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
-using osu.Game.Beatmaps;
-using osu.Game.Beatmaps.ControlPoints;
 using osu.Game.Rulesets.Mania.Objects.Types;
 using osu.Game.Rulesets.Objects;
 
@@ -12,12 +10,6 @@ namespace osu.Game.Rulesets.Mania.Objects
     {
         public virtual int Column { get; set; }
 
-        protected override void ApplyDefaultsToSelf(ControlPointInfo controlPointInfo, BeatmapDifficulty difficulty)
-        {
-            base.ApplyDefaultsToSelf(controlPointInfo, difficulty);
-
-            HitWindows.AllowsPerfect = true;
-            HitWindows.AllowsOk = true;
-        }
+        protected override HitWindows CreateHitWindows() => new ManiaHitWindows();
     }
 }
