+++ osu.Game/Rulesets/Objects/Legacy/Mania/ConvertHit.cs
@@ -14,6 +14,6 @@ namespace osu.Game.Rulesets.Objects.Legacy.Mania
 
         public bool NewCombo { get; set; }
 
-        protected override HitWindows CreateHitWindows() => new ConvertHitWindows();
+        protected override HitWindows CreateHitWindows() => null;
     }
 }
