+++ osu.Game/Rulesets/Objects/Legacy/Mania/ConvertHold.cs
@@ -13,6 +13,6 @@ namespace osu.Game.Rulesets.Objects.Legacy.Mania
 
         public double Duration => EndTime - StartTime;
 
-        protected override HitWindows CreateHitWindows() => new ConvertHitWindows();
+        protected override HitWindows CreateHitWindows() => null;
     }
 }
