+++ osu.Game/Rulesets/Objects/Legacy/Mania/ConvertSpinner.cs
@@ -16,6 +16,6 @@ namespace osu.Game.Rulesets.Objects.Legacy.Mania
 
         public float X { get; set; }
 
-        protected override HitWindows CreateHitWindows() => new ConvertHitWindows();
+        protected override HitWindows CreateHitWindows() => null;
     }
 }
