+++ osu.Game/Rulesets/Objects/Legacy/Taiko/ConvertSlider.cs
@@ -12,6 +12,6 @@ namespace osu.Game.Rulesets.Objects.Legacy.Taiko
     {
         public bool NewCombo { get; set; }
 
-        protected override HitWindows CreateHitWindows() => new ConvertHitWindows();
+        protected override HitWindows CreateHitWindows() => null;
     }
 }
